requirements.txt
absl-py==2.1.0
accimage @ file:///home/conda/feedstock_root/build_artifacts/accimage_1636476522462/work
ADEval==1.1.0
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
Brotli @ file:///croot/brotli-split_1714483155106/work
bs4==0.0.2
cachetools==5.4.0
certifi @ file:///croot/certifi_1725551672989/work/certifi
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.7
conda-pack @ file:///croot/conda-pack_1710258022278/work
contourpy==1.1.1
cycler==0.12.1
decorator==5.1.1
einops==0.8.0
et-xmlfile==1.1.0
executing==2.0.1
faiss-gpu==1.7.2
fastprogress==1.0.3
filelock @ file:///croot/filelock_1700591183607/work
fonttools==4.53.1
FrEIA==0.2
fsspec==2024.6.1
future==1.0.0
fvcore==0.1.5.post20221221
geomloss==0.2.6
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645455532332/work
google-auth==2.32.0
google-auth-oauthlib==1.0.0
grpcio==1.65.2
html5lib==1.1
huggingface-hub==0.24.5
idna @ file:///croot/idna_1714398848350/work
imageio==2.34.2
imgaug==0.4.0
importlib_metadata==8.2.0
importlib_resources==6.4.0
iopath==0.1.10
ipdb==0.13.13
ipython==8.12.3
jedi==0.19.1
Jinja2 @ file:///croot/jinja2_1716993405101/work
joblib==1.4.2
kiwisolver==1.4.5
lazy_loader==0.4
llvmlite==0.41.1
lxml==5.3.0
mamba-ssm==2.2.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe @ file:///croot/markupsafe_1704205993651/work
matplotlib==3.7.5
matplotlib-inline==0.1.7
mdurl==0.1.2
memory-profiler==0.61.0
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
mmcls==1.0.0rc6
mmdet==2.25.3
mmselfsup==1.0.0
model-index==0.1.11
modelindex==0.0.2
mpmath @ file:///croot/mpmath_1690848262763/work
networkx @ file:///croot/networkx_1690561992265/work
ninja==1.11.1.1
numba==0.58.1
numpy @ file:///work/mkl/numpy_and_numpy_base_1682953417311/work
numpy-hilbert-curve==1.0.1
oauthlib==3.2.2
opencv-python==4.10.0.84
openpyxl==3.1.5
ordered-set==4.1.0
packaging==24.1
pandas==2.0.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow @ file:///croot/pillow_1721059439630/work
portalocker==2.10.1
prompt_toolkit==3.0.47
protobuf==3.20.1
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycocotools==2.0.7
Pygments==2.18.0
pyparsing==3.1.2
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.9.0.post0
pytz==2024.1
PyWavelets==1.4.1
PyYAML @ file:///croot/pyyaml_1698096049011/work
pyzorder==0.0.2
regex==2024.7.24
requests @ file:///croot/requests_1721410876868/work
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
safetensors==0.4.3
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
shapely==2.0.5
six==1.16.0
soupsieve==2.6
stack-data==0.6.3
sympy @ file:///croot/sympy_1701397643339/work
tabulate==0.9.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.4.0
terminaltables==3.1.10
threadpoolctl==3.5.0
tifffile==2023.7.10
timm==0.8.15.dev0
tokenizers==0.19.1
tomli==2.0.1
torch==2.1.2
torch-tb-profiler==0.4.3
torchaudio==2.1.2
torchvision==0.16.2
tqdm==4.66.4
traitlets==5.14.3
transformers==4.43.3
triton==2.1.0
typing_extensions @ file:///croot/typing_extensions_1715268824938/work
tzdata==2024.1
urllib3 @ file:///croot/urllib3_1718912636303/work
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.3
yacs==0.1.8
zipp==3.19.2
